+++ intelmq/tests/bots/parsers/microsoft/test_parser_ctip.py
@@ -16,8 +16,6 @@ EXAMPLE_PARSED = json.loads(EXAMPLE_DATA)
 EXAMPLE_REPORT = {
     "__type": "Report",
     "feed.accuracy": 100.0,
-    "feed.name": "ctip",
-    "feed.url": "http://192.168.56.102/",
     "time.observation": "2016-06-15T09:25:26+00:00",
     "raw": base64_encode(EXAMPLE_DATA)
 }
@@ -32,8 +30,6 @@ EXAMPLE_EVENTS = [{
     'extra.isproductlicensed': 'true',
     'tlp': 'GREEN',
     'feed.accuracy': 20.0,
-    'feed.name': 'ctip',
-    'feed.url': 'http://192.168.56.102/',
     'malware.name': 'b85-r2s',
     'source.asn': 65536,
     'source.ip': '224.0.5.8',
@@ -50,8 +46,6 @@ EXAMPLE_EVENTS = [{
     'extra.isproductlicensed': 'true',
     'tlp': 'GREEN',
     'feed.accuracy': 100.0,
-    'feed.name': 'ctip',
-    'feed.url': 'http://192.168.56.102/',
     'malware.name': 'conficker',
     'source.asn': 64496,
     'source.ip': '10.0.0.5',
@@ -68,8 +62,6 @@ EXAMPLE_EVENTS = [{
     'extra.isproductlicensed': 'true',
     'tlp': 'GREEN',
     'feed.accuracy': 100.0,
-    'feed.name': 'ctip',
-    'feed.url': 'http://192.168.56.102/',
     'malware.name': 'b106-jenxcus',
     'source.asn': 64511,
     'source.ip': '19.168.46.126',
@@ -86,33 +78,30 @@ EXAMPLE_EVENTS = [{
     'extra.isproductlicensed': 'true',
     'tlp': 'GREEN',
     'feed.accuracy': 100.0,
-    'feed.name': 'ctip',
-    'feed.url': 'http://192.168.56.102/',
     'malware.name': 'b75-s12',
     'source.asn': 65536,
     'source.ip': '198.51.100.100',
     'source.port': 42996,
     'time.source': '2018-02-06T09:43:19+00:00',
     "raw": base64_encode(json.dumps([EXAMPLE_PARSED[3]], sort_keys=True)),
-    }, {
+    }, {  # ignore hostname if invalid
     "__type": "Event",
     'classification.type': 'infected system',
     'destination.ip': '192.88.99.209',
     'destination.port': 16465,
     'event_description.text': 'Host is a member of Botnet B68-2-64',
-    'extra.additionalmetadata': [r'any??thing\x0000can\x01be!here??'],
+    "extra.additionalmetadata": ["??\f??????9??\u001e??P??????%??a????????.??????\u0016/(????\u0006??$s????v????zU??,??????#_??d??\u0013????????????\u00131!}\u001e??2/+????@??P1??\u0016??????????<>tR??}^??t-h??G??F??+????l"],
     'extra.ispartnershareable': 'true',
     'extra.isproductlicensed': 'true',
+    "extra.user_agent": "Q\u000bZk??8\u001d????????p??z????z??q)??\u001e\u0012??'??????,i3??????\u0006??q??\u0007d=????????-??b????={????????X??{iN}??\u0017\u0011\u0010|q??\u0014",
     'tlp': 'GREEN',
     'feed.accuracy': 100.0,
-    'feed.name': 'ctip',
-    'feed.url': 'http://192.168.56.102/',
-    'malware.name': 'b75-s12',
+    'malware.name': 'b68-2-64',
     'source.asn': 64504,
     'source.ip': '224.34.234.52',
     'source.port': 55522,
-    'time.source': '2019-03-17T09:05:50+00"00',
-    "raw": base64_encode(json.dumps([EXAMPLE_PARSED[3]], sort_keys=True)),
+    'time.source': '2019-03-17T09:05:50+00:00',
+    "raw": base64_encode(json.dumps([EXAMPLE_PARSED[4]], sort_keys=True)),
     },
     ]
 
@@ -130,7 +119,7 @@ class TestMicrosoftCTIPParserBot(test.BotTestCase, unittest.TestCase):
     def test_event(self):
         """ Test if correct Events have been produced. """
         self.run_bot()
-        for i in range(4):
+        for i in range(5):
             self.assertMessageEqual(i, EXAMPLE_EVENTS[i])